Move App and WidgetDataProvider components to Components folder
